refactor: extract BookDetailSubviews and enhance MVVM state synchronization
- Extracted subviews from BookDetailView.swift into BookDetailSubviews.swift.
- Moved UI logic (readBook, cacheFormat, clearFormat, etc.) from BookDetailSubviews into BookDetailViewModel to establish a clean MVVM boundary.
- Fixed UI state synchronization delays by updating book.lastUpdated = Date() in ModelData.clearCache and CalibreServerService.handleLibraryBookOne.
- Removed dead presentation bindings from BookDetailViewModel and accurately mapped them to View-level local states.
- Added comprehensive unit tests and UI tests for BookDetailViewModel and fixed failing assertions.
- Reverted temporary booksInShelf lookup workaround in ModelData.convert as UI lag is now properly handled by Realm notifications.
