+++ .agents/rules/workflow-rules.md
@@ -3,7 +3,11 @@ trigger: always_on
 description: My custom workflow rules
 ---
 
-在开始任何任务之前，必须先读取 memory-bank 中的文档了解上下文；完成任务后，自动更新 memory-bank 文件
+在开始任何任务之前，必须先读取 .agents/memory-bank 中的文档了解上下文；完成任务后，自动更新 .agents/memory-bank 文件
 
-始终通过mcp修改项目文件
+每次先生成研究报告并制定实施计划，提请用户确认
+
+始终通过xcode mcp修改工程内容
+
+始终通过xcode build 和 test进行验证
