+++ YetAnotherEBookReaderTests/YetAnotherEBookReaderTests.swift
@@ -1,0 +1,8 @@
+import XCTest
+@testable import YetAnotherEBookReader
+
+class YetAnotherEBookReaderTests: XCTestCase {
+    func testExample() throws {
+        XCTAssertTrue(true)
+    }
+}
