+++ YetAnotherEBookReaderUITests/YetAnotherEBookReaderUITests.swift
@@ -1,0 +1,8 @@
+import XCTest
+
+class YetAnotherEBookReaderUITests: XCTestCase {
+    func testExample() throws {
+        let app = XCUIApplication()
+        app.launch()
+    }
+}
